pyproject.toml
[tool.poetry]
name = "opeanaiassistant"
version = "0.1.0"
description = "An example application with OpenAI Assistant and a vector store. UI built with Streamlit."
authors = ["Bartosz Mikulski <blog@mikulskibartosz.name>"]
license = "Apache License 2.0"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "1.34.0"
openai = "1.26.0"
python-dotenv = "1.0.1"
youtube-transcript-api = "0.6.2"

[tool.poetry.dev-dependencies]
black = "24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
